refactor: move sector types to common
