+++ storage-provider/common/Cargo.toml
@@ -13,6 +13,7 @@ primitives = { workspace = true }
 storagext = { workspace = true, features = ["clap"] }
 
 chrono = { workspace = true, features = ["serde"] }
+codec = { workspace = true }
 cid = { workspace = true, features = ["serde", "std"] }
 filecoin-hashers.workspace = true
 fr32.workspace = true
@@ -22,9 +23,11 @@ serde_json = { workspace = true }
 storage-proofs-core.workspace = true
 subxt = { workspace = true }
 thiserror = { workspace = true }
-
-[dev-dependencies]
-polka-storage-proofs = { workspace = true, features = ["std"] }
+polka-storage-proofs = { workspace = true, features = ["std", "substrate"] }
+tokio = { workspace = true, features = ["fs"] }
+tokio-util = { workspace = true, features = ["rt"] }
+tracing = { workspace = true }
+hex = { workspace = true, features = ["std"] }
 
 [lints]
 workspace = true
